requirements.txt
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dotenv==0.20.0
